+++ xbmc/utils/i18n/Iso639.cpp
@@ -6,7 +6,7 @@
  *  See LICENSES/README.md for more information.
  */
 
-#include "utils/i18n/TableISO639.h"
+#include "utils/i18n/Iso639.h"
 
 #include <algorithm>
 #include <cstdint>
