+++ xbmc/utils/i18n/Iso639_2.cpp
@@ -9,8 +9,8 @@
 #include "utils/i18n/Iso639_2.h"
 
 #include "utils/StringUtils.h"
+#include "utils/i18n/Iso639.h"
 #include "utils/i18n/Iso639_2_Table.h"
-#include "utils/i18n/TableISO639.h"
 
 #include <algorithm>
